BHA-225 | Change internalId type to Long to accommodate larger number of graph nodes | Jayanta

+++ src/main/scala/com/bharatsim/engine/graph/GraphProvider.scala
@@ -246,5 +246,5 @@ trait GraphProvider {
 }
 
 object GraphProvider {
-  type NodeId = Int
+  type NodeId = Long
 }
